Properly sort conferences.
This fixes issue #159. The problem was that the sort function was returning the wrong values. I had originally assumed it was an "is ordered before" test, when it is actually a comparison function and wanted a positive, negative, or zero value to indicate relative ordering.

I've verified that the updated code properly sorts the conferences in both Safari and Chrome.