Shift end dates forward 36 hours
iCalendar format is a bit strange with how it deals with dates. A date in a field is that starting _instant_ of when a thing starts or ends. Therefore an event that starts at `2020-01-22` and ends at `2020-01-22` has a length of "0". Conventionally we think it would mean "one day", but that's not how iCalendar format works. Therefore, we need to shift the end date _forward_ a day to represent the first instant of _after_ the event ends. Or in other words, an all-day event on Jan 22 starts at `2020-01-22` and ends at `2020-01-23`.

However, that's not *quite* good enough, because we do have rogue leap hours that mix in. Therefore, we're going to shift forward by 36 hours so we can be sure we're solidly during the "next day" before we try to format it back down to just the year-month-day.